# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  sku: string | null;
  price: number | null;
  categoryId: number | null;
  netsuiteId: string | null;
  imageUrl: string | null;
}